Works for creating trust
