+++ src/locales/cn/dashboard.ts
@@ -90,9 +90,8 @@ export default {
   'label.asset-list.unclaimed': '未领',
   'label.asset-list.total-amount': '总额',
   'asset-list.title': '我的信托',
-  'create-new-trust.menu.step-one': '请选择资产',
-  'create-new-trust.menu.step-two': '请设置信托规则',
-  'create-new-trust.menu.step-three': '完成信托并确认',
+  'create-new-trust.menu.step-one': '请设置信托规则',
+  'create-new-trust.menu.step-two': '完成信托并确认',
   'create-new-trust.fund-deposit': '我想从此钱包里充值',
   'create-new-trust.fund-deposit-asset': '我想添加此类资产',
   'create-new-trust.is-trust-revocable-or-not': '信托是否可撤销？',
