+++ src/locales/en/dashboard.ts
@@ -99,9 +99,8 @@ export default {
   'label.asset-list.unclaimed': 'Unclaimed',
   'label.asset-list.total-amount': 'Total Amount',
   'asset-list.title': 'Your assets in trusts',
-  'create-new-trust.menu.step-one': 'Choose source of funds',
-  'create-new-trust.menu.step-two': 'Enter trust details',
-  'create-new-trust.menu.step-three': 'Deploy and confirm',
+  'create-new-trust.menu.step-one': 'Enter trust details',
+  'create-new-trust.menu.step-two': 'Deploy and confirm',
   'create-new-trust.fund-deposit': 'I want to deposit funds from',
   'create-new-trust.fund-deposit-asset': 'I want to deposit this asset:',
   'create-new-trust.is-trust-revocable-or-not': 'Is the trust revocable or not?',
